pyproject.toml
[tool.poetry]
name = "oic"
version = "0.1.0"
description = ""
authors = ["Stefan Bethke <stb@lassitu.de>"]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
tornado = "^6.2"
asyncio-mqtt = "^0.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
